Deserialize EntryHeader parameters
The EntryHeader would set the map of all its parameters as its name
member, which semantically does not make sense and requires a user to
parse the name member itself to find the actual name of the header.

Deserialize the EntryHeader according to the TSP to set members
"name", "tooltip" and "data_type".