add representation methods and encoders
The encoders are useful to store objects on disk between multiple
calls to a program. This is useful in some cases where you need
to recover the previous state of your run.

The representations implementation is very useful to print the
results of the requests to know what you got. It is very useful
for debugging and testing out things when developping.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>